refactor: cleanup widget and blog styles
