+++ example/content/posts/2019-10-27--hello-world/index.mdx
@@ -1,10 +1,14 @@
 ---
-slug: '/posts/hello-world'
+slug: /posts/hello-world
 date: '2019-10-27T09:00:00.000Z'
 title: Hello World
-description: 'This is the Hello World example post document ...'
+description: >
+  This is the Hello World example post document ...
+excerpt: >
+  Introduction article for this this. Start here to learn more about this
+  theme and the available shortcodes, styles, and features.
 keywords: ['react']
-banner: ''
+banner: https://res.cloudinary.com/chrisvogt/image/upload/v1573540598/photos/2019-nyc-world-pride/IMG_7034.jpg
 ---
 
 import { HelloWorld } from './hello-world'
