+++ example/content/posts/yet-another-page.mdx
@@ -1,0 +1,10 @@
+---
+slug: '/posts/yet-another-page'
+date: '2019-12-25T09:00:00.000Z'
+title: Yet Another Page
+description: 'This is the Hello World example post document ...'
+keywords: ['react']
+banner: ''
+---
+
+Lorum ipsum
