+++ theme/src/components/widgets/recent-posts/index.js
@@ -1,0 +1,3 @@
+import RecentPostsWidget from './recent-posts-widget'
+
+export default RecentPostsWidget
