+++ theme/src/components/widgets/widget.js
@@ -2,27 +2,9 @@
 import { jsx } from 'theme-ui'
 import PropTypes from 'prop-types'
 
-import theme from '../../gatsby-plugin-theme-ui'
-
 const Widget = ({ children, id }) => {
   return (
-    <div
-      sx={{
-        backgroundColor: [`background`, `#fdfdfd`],
-        borderLeftWidth: [``, `3px`],
-        borderLeftStyle: [`none`, `solid`],
-        borderLeftColor: [``, `#9231A7`],
-        borderRadius: [``, `2px`],
-        borderTop: [``, `1px solid white`],
-        boxShadow: [``, theme.shadows.default],
-        ...(id ? { id } : {}),
-        mb: [3, 4],
-        p: [0, 3, 4],
-        '&:not(:last-of-type)': {
-          borderBottom: [`2px solid #f9f9f9`, `1px solid white`]
-        }
-      }}
-    >
+    <div sx={{ variant: 'styles.Widget' }} {...(id ? { id } : {})}>
       {children}
     </div>
   )
@@ -31,7 +13,7 @@ const Widget = ({ children, id }) => {
 Widget.propTypes = {
   /** The elements to render within the widget. */
   children: PropTypes.node.isRequired,
-  /** An id added to the widget wrapper. */
+  /** An id added to the widget wrapper element. */
   id: PropTypes.string
 }
